move compaction client to compaction runner crate.
