+++ Cargo.toml
@@ -34,7 +34,6 @@ members = [
     "src/catalog_impls",
     "src/cluster",
     "src/common_types", 
-    "src/compaction_client",
     "src/components/alloc_tracker",
     "src/components/arena",
     "src/components/arrow_ext",
@@ -111,7 +110,6 @@ cluster = { path = "src/cluster" }
 criterion = "0.5"
 horaedb-client = "1.0.2"
 common_types = { path = "src/common_types" }
-compaction_client = { path = "src/compaction_client" }
 datafusion = { git = "https://github.com/CeresDB/arrow-datafusion.git", rev = "e21b03154" }
 datafusion-proto = { git = "https://github.com/CeresDB/arrow-datafusion.git", rev = "e21b03154" }
 derive_builder = "0.12"
