+++ src/analytic_engine/Cargo.toml
@@ -52,7 +52,6 @@ bytes_ext = { workspace = true }
 cluster = { workspace = true }
 codec = { workspace = true }
 common_types = { workspace = true }
-compaction_client = { workspace = true }
 datafusion = { workspace = true }
 future_ext = { workspace = true }
 futures = { workspace = true }
@@ -76,10 +75,12 @@ parquet_ext = { workspace = true }
 prometheus = { workspace = true }
 prost = { workspace = true }
 remote_engine_client = { workspace = true }
+reqwest = { workspace = true }
 router = { workspace = true }
 runtime = { workspace = true }
 sampling_cache = { workspace = true }
 serde = { workspace = true }
+serde_json = { workspace = true }
 size_ext = { workspace = true }
 skiplist = { path = "../components/skiplist" }
 smallvec = { workspace = true }
@@ -90,7 +91,9 @@ tempfile = { workspace = true, optional = true }
 thiserror = { workspace = true }
 time_ext = { workspace = true }
 tokio = { workspace = true }
+tonic = { workspace = true }
 trace_metric = { workspace = true }
+url = "2.2"
 wal = { workspace = true }
 xorfilter-rs = { workspace = true }
