+++ src/horaedb/Cargo.toml
@@ -44,7 +44,6 @@ catalog_impls     = { workspace = true }
 clap              = { workspace = true }
 cluster           = { workspace = true }
 common_types      = { workspace = true }
-compaction_client = { workspace = true }
 datafusion        = { workspace = true }
 df_operator       = { workspace = true }
 etcd-client       = { workspace = true }
