+++ src/router/Cargo.toml
@@ -34,7 +34,6 @@ workspace = true
 async-trait = { workspace = true }
 cluster = { workspace = true }
 common_types = { workspace = true }
-compaction_client = { workspace = true }
 generic_error = { workspace = true }
 horaedbproto = { workspace = true }
 logger = { workspace = true }
